renamed score_calculator.rs
