+++ crates/driver/src/domain/competition/solution/solver_score.rs
